moved display functions into their own file
