+++ src/lib.rs
@@ -1,18 +1,19 @@
 #![no_std]
 
-extern crate alloc;
-
-use numworks::display::{self, CHAR_HEIGHT, Color, LARGE_CHAR_WIDTH, SCREEN_HEIGHT, SCREEN_WIDTH};
+use numworks::display::{
+    self, CHAR_HEIGHT, EadkColor, LARGE_CHAR_WIDTH, SCREEN_HEIGHT, SCREEN_WIDTH,
+};
 
 /// The core of the application logic
 pub fn main() {
     const MESSAGE: &str = "Hey !!";
-    display::clear(Color::GREEN);
-    display::draw_string_large(
+    display::clear(EadkColor::GREEN);
+    display::draw_string(
         MESSAGE,
         (SCREEN_WIDTH - (MESSAGE.len() as u16 * LARGE_CHAR_WIDTH)) / 2,
         (SCREEN_HEIGHT - CHAR_HEIGHT) / 2,
-        Color::BLACK,
-        Color::WHITE,
+        true,
+        EadkColor::BLACK,
+        EadkColor::WHITE,
     );
 }
